pyproject.toml
[project]
name = "contributor-network"
version = "0.1.0"
description = "A visualization of contributors to projects"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.9",
    "pygithub>=2.8.1",
]

[project.scripts]
contributor-network = "contributor_network.cli:main"

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]

[tool.mypy]
files = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
